Revert DAGCircuit stuff.
