+++ crates/circuit/Cargo.toml
@@ -12,4 +12,3 @@ doctest = false
 [dependencies]
 hashbrown.workspace = true
 pyo3.workspace = true
-rustworkx-core = "0.14"
