+++ crates/circuit/src/circuit_data.rs
@@ -10,7 +10,6 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-use crate::bit_packing::{pack, unpack, BitAsKey, PackedInstruction};
 use crate::circuit_instruction::CircuitInstruction;
 use crate::intern_context::{BitType, IndexType, InternContext};
 use crate::SliceOrInt;
@@ -21,6 +20,66 @@ use pyo3::prelude::*;
 use pyo3::types::{PyList, PySet, PySlice, PyTuple, PyType};
 use pyo3::{PyObject, PyResult, PyTraverseError, PyVisit};
 use std::hash::{Hash, Hasher};
+
+/// Private type used to store instructions with interned arg lists.
+#[derive(Clone, Debug)]
+struct PackedInstruction {
+    /// The Python-side operation instance.
+    op: PyObject,
+    /// The index under which the interner has stored `qubits`.
+    qubits_id: IndexType,
+    /// The index under which the interner has stored `clbits`.
+    clbits_id: IndexType,
+}
+
+/// Private wrapper for Python-side Bit instances that implements
+/// [Hash] and [Eq], allowing them to be used in Rust hash-based
+/// sets and maps.
+///
+/// Python's `hash()` is called on the wrapped Bit instance during
+/// construction and returned from Rust's [Hash] trait impl.
+/// The impl of [PartialEq] first compares the native Py pointers
+/// to determine equality. If these are not equal, only then does
+/// it call `repr()` on both sides, which has a significant
+/// performance advantage.
+#[derive(Clone, Debug)]
+struct BitAsKey {
+    /// Python's `hash()` of the wrapped instance.
+    hash: isize,
+    /// The wrapped instance.
+    bit: PyObject,
+}
+
+impl BitAsKey {
+    fn new(bit: &Bound<PyAny>) -> PyResult<Self> {
+        Ok(BitAsKey {
+            hash: bit.hash()?,
+            bit: bit.into_py(bit.py()),
+        })
+    }
+}
+
+impl Hash for BitAsKey {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        state.write_isize(self.hash);
+    }
+}
+
+impl PartialEq for BitAsKey {
+    fn eq(&self, other: &Self) -> bool {
+        self.bit.is(&other.bit)
+            || Python::with_gil(|py| {
+                self.bit
+                    .bind(py)
+                    .repr()
+                    .unwrap()
+                    .eq(other.bit.bind(py).repr().unwrap())
+                    .unwrap()
+            })
+    }
+}
+
+impl Eq for BitAsKey {}
 
 /// A container for :class:`.QuantumCircuit` instruction listings that stores
 /// :class:`.CircuitInstruction` instances in a packed form by interning
@@ -265,7 +324,7 @@ impl CircuitData {
             0,
         )?;
         res.intern_context = self.intern_context.clone();
-        res.data = self.data.clone();
+        res.data.clone_from(&self.data);
         Ok(res)
     }
 
@@ -400,21 +459,14 @@ impl CircuitData {
         clbits: Option<&Bound<PyAny>>,
     ) -> PyResult<()> {
         let mut temp = CircuitData::new(py, qubits, clbits, None, 0)?;
-        if temp.qubits_native.len() < self.qubits_native.len() {
-            return Err(PyValueError::new_err(format!(
-                "Replacement 'qubits' of size {:?} must contain at least {:?} bits.",
-                temp.qubits_native.len(),
-                self.qubits_native.len(),
-            )));
-        }
-        if temp.clbits_native.len() < self.clbits_native.len() {
-            return Err(PyValueError::new_err(format!(
-                "Replacement 'clbits' of size {:?} must contain at least {:?} bits.",
-                temp.clbits_native.len(),
-                self.clbits_native.len(),
-            )));
-        }
         if qubits.is_some() {
+            if temp.qubits_native.len() < self.qubits_native.len() {
+                return Err(PyValueError::new_err(format!(
+                    "Replacement 'qubits' of size {:?} must contain at least {:?} bits.",
+                    temp.qubits_native.len(),
+                    self.qubits_native.len(),
+                )));
+            }
             std::mem::swap(&mut temp.qubits, &mut self.qubits);
             std::mem::swap(&mut temp.qubits_native, &mut self.qubits_native);
             std::mem::swap(
@@ -423,6 +475,13 @@ impl CircuitData {
             );
         }
         if clbits.is_some() {
+            if temp.clbits_native.len() < self.clbits_native.len() {
+                return Err(PyValueError::new_err(format!(
+                    "Replacement 'clbits' of size {:?} must contain at least {:?} bits.",
+                    temp.clbits_native.len(),
+                    self.clbits_native.len(),
+                )));
+            }
             std::mem::swap(&mut temp.clbits, &mut self.clbits);
             std::mem::swap(&mut temp.clbits_native, &mut self.clbits_native);
             std::mem::swap(
@@ -736,29 +795,61 @@ impl CircuitData {
         Ok(index as usize)
     }
 
-    #[inline]
+    /// Returns a [PackedInstruction] containing the original operation
+    /// of `elem` and [InternContext] indices of its `qubits` and `clbits`
+    /// fields.
     fn pack(
         &mut self,
         py: Python<'_>,
         inst: PyRef<CircuitInstruction>,
     ) -> PyResult<PackedInstruction> {
-        pack(
-            py,
-            &mut self.intern_context,
-            &self.qubit_indices_native,
-            &self.clbit_indices_native,
-            inst,
-        )
+        let mut interned_bits =
+            |indices: &HashMap<BitAsKey, BitType>, bits: &Bound<PyTuple>| -> PyResult<IndexType> {
+                let args = bits
+                    .into_iter()
+                    .map(|b| {
+                        let key = BitAsKey::new(&b)?;
+                        indices.get(&key).copied().ok_or_else(|| {
+                            PyKeyError::new_err(format!(
+                                "Bit {:?} has not been added to this circuit.",
+                                b
+                            ))
+                        })
+                    })
+                    .collect::<PyResult<Vec<BitType>>>()?;
+                self.intern_context.intern(args)
+            };
+        Ok(PackedInstruction {
+            op: inst.operation.clone_ref(py),
+            qubits_id: interned_bits(&self.qubit_indices_native, inst.qubits.bind(py))?,
+            clbits_id: interned_bits(&self.clbit_indices_native, inst.clbits.bind(py))?,
+        })
     }
 
-    #[inline]
     fn unpack(&self, py: Python<'_>, inst: &PackedInstruction) -> PyResult<Py<CircuitInstruction>> {
-        unpack(
+        Py::new(
             py,
-            &self.intern_context,
-            &self.qubits_native,
-            &self.clbits_native,
-            inst,
+            CircuitInstruction {
+                operation: inst.op.clone_ref(py),
+                qubits: PyTuple::new_bound(
+                    py,
+                    self.intern_context
+                        .lookup(inst.qubits_id)
+                        .iter()
+                        .map(|i| self.qubits_native[*i as usize].clone_ref(py))
+                        .collect::<Vec<_>>(),
+                )
+                .unbind(),
+                clbits: PyTuple::new_bound(
+                    py,
+                    self.intern_context
+                        .lookup(inst.clbits_id)
+                        .iter()
+                        .map(|i| self.clbits_native[*i as usize].clone_ref(py))
+                        .collect::<Vec<_>>(),
+                )
+                .unbind(),
+            },
         )
     }
 }
